Win Video Capture: support HDYC pixel format.
This is a reland of a previous crrev.com/317073010
that caused instabilities with Third party drivers,
probably because of some subtle changes in
enumeration. This CL lands only the new pixel format.

Original description -------------------------------------

This pixel format is a synonym of YUYV and is needed
for at least the DeckLink video capture card.
This format is not specified in Windows headers
probably because this is just a synonym of other
pixel formats, IOW they are all the same except for
the first word, which is the name read in hex
backwards:

30323449 --> 0x30 0x32 0x34 0x49 -> 024I
43594448 --> 0x43 0x59 0x44 0x48 -> CYDH

and then all the YUV (422 and 420 alike) use the same
trailing sequence
... 0x0000, 0x0010, {0x80, 0x00, 0x00, 0xaa, 0x00, 0x38, 0x9b, 0x71}}

HDYC: http://www.fourcc.org/yuv.php#HDYC


BUG=380939

Review URL: https://codereview.chromium.org/348033004