MacViews: Add an NSWindowDelegate, initially for NativeWidgetPrivate::Get* static functions.
The NSWindow* -> NativeWidgetMac* mapping required to implement
GetNativeWidgetForNativeWindow() is achieved by introducing an
NSWindowDelegate that is attached to NSWindow objects when they are
bridged by BridgedNativeWidget.

This allows the following tests to pass:
- NativeWidgetTest.CreateNativeWidget
- NativeWidgetTest.GetNativeWidgetForNativeView
- NativeWidgetTest.GetTopLevelNativeWidget1

NativeWidgetTest.GetTopLevelNativeWidget2 will currently crash on
MacViews because NativeViewHostMac is not implemented.

The NSWindowDelegate will be used for a lot more later.

BUG=378134
TEST=views_unittests

Review URL: https://codereview.chromium.org/346773002