allow autorotate when converting from pdf to emf
I'm not sure if this is the "right" fix, but the behaviour this way
is better than currently. The pages will autorotate to fit, but they'll
end up being (sideways) portrait pages when printed to a pdf/oxps/etc.

This seems to be what PWGRasterConverter and Cloud Print do to handle
this case.

(This will probably need to be merged to 37 branch)

R=vitalybuka@chromium.org
BUG=387309

Review URL: https://codereview.chromium.org/355443003