Mac ÜC: Fix bugs encountered by presubmits
Pipe latency info through the BrowerCompositorViewMac, so that the gpu
tests that use it for screenshots pass. Do this by adding an interface
BrowserCompositorViewClient, which informs the RenderWidgetHostViewMac
when the content is available for screenshot.

Remove the calls to -[BrowserCompositorViewMac layoutLayers], since
they don't do anything anymore (layers are correctly position by default
now that we use a flipped coordinate system).

Make -[SoftwareLayerMac setContentsToData] set the layer bounds of
the software layer directly, rather than doing that in layoutLayers.

Disable attempts at doing a smooth resize when there exists a browser
compositor, as that feature is not yet implemented.

BUG=314190

Review URL: https://codereview.chromium.org/337233004