Revert 278393 "Report when blacklist.cc fails to setup the black..."
Reverting on beta branch as systemk.dll crashes have spiked dramatically in recent dev push

https://crash.corp.google.com/browse?q=product.name%3D%27Chrome%27%20AND%20custom_data.ChromeCrashProto.ptype%3D%27browser%27%20AND%20custom_data.ChromeCrashProto.magic_signature_1.name%3D%27Third%20party%27%20AND%20custom_data.ChromeCrashProto.magic_signature_2.name%3D%27%5BThird%20party%5D%20%25ProgramFiles%25%5C%5Csettings%20manager%5C%5Csystemk%5C%5Csystemk.dll%27%20AND%20product.version%3D%2737.0.2062.0%27

We'd like to keep this DLL blacklisted, however it appears as if it is creating a crash on startup.
If we can mitigate this, we'd be happy to re-incorporate this change.

> Report when blacklist.cc fails to setup the blacklisting thunk.
> 
> BUG=
> 
> Review URL: https://codereview.chromium.org/343613003

TBR=csharp@chromium.org

Review URL: https://codereview.chromium.org/356593003