Separate out parts of chrome_browser.gypi.
The separate file lists can be shared with the GN build, so are moved to the top of the file in a variables section and then referenced in the conditions when appropriate.

This doesn't move everything, in particular there are still a lot of ChromeOS, Android, iOS, and a few features with exclusion lists.

BUG=
R=thestig@chromium.org

Review URL: https://codereview.chromium.org/336513008