[Cast] A really simple skeleton mirror session simulation executable.
The program takes arguments simulation run id (to be used as tags) and
a file path and writes out a line to the given file path.

The end goal is to also provide simulation parameters and then have the
program simulate a session using the parameters, and at the end write
out the raw event log to the given path with the given sim run id tag.

Review URL: https://codereview.chromium.org/355523002