Reduce boilerplate in PasswordFormConversionUtilsTest.
Introduce a PasswordFormBuilder so that we can describe testing forms in a more concise and clearer way (instead of using raw HTML in string literals).

This is not critical right now, but will be very useful once we have a lot more tests for autocomplete attributes.

BUG=375333

Review URL: https://codereview.chromium.org/336403003