Added --custom-launcher-page. Adds a page to the experimental launcher.
This allows the user to specify a custom web page (which must be in the
chrome-extension:// scheme) which will appear as a third top-level page
in the experimental launcher, alongside the start and apps pages.

This command-line switch is intended to be temporary; eventually,
extensions will be able to specify their own custom launcher pages. For
now, the command-line switch is sufficient for testing.

Most of the Chrome APIs are not yet available for these pages.

BUG=386004
TBR=mukai@chromium.org,jamescook@chromium.org

Review URL: https://codereview.chromium.org/352033002