Revert of Add PRESUBMIT.py warning for contradictory NOTREACHED() use. (https://codereview.chromium.org/344563003/)
Reason for revert:
It seems to be checking code that is not part of the diff. 

I also think there are some legitimate cases for not reached in the middle of a block but that is something that can be discussed on the bug.

Filed https://code.google.com/p/chromium/issues/detail?id=388731 to track.

Original issue's description:
> Add PRESUBMIT.py warning for contradictory NOTREACHED() use.
> 
> As recently discussed on chromium-dev [1] and subsequently incorporated into the
> Chromium Coding Style [2], NOTREACHED() followed by code is an anti-pattern.
> This CL lets PRESUBMIT.py detect most of these cases and print a warning to the
> developer.
> [1] https://groups.google.com/a/chromium.org/d/msg/chromium-dev/c2X0D1Z6X2o/ddkFLqQP0P0J
> [2] http://www.chromium.org/developers/coding-style#TOC-CHECK-DCHECK-and-NOTREACHED-
> 
> BUG=none
> 
> Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=279681

TBR=maruel@chromium.org,tnagel@chromium.org
NOTREECHECKS=true
NOTRY=true
BUG=none

Review URL: https://codereview.chromium.org/357693004