Ignore Manager.Services list in ShillPropertyHandler
We were calling UpdateObserved() for both Manager.Services and
ServiceCompleteList which was incorrect.

It turns out that Shill may push the Service.Visible property before
adding a service to Manager.Services, so if we only observer entries in
Manager.Services we will miss the update.

The easiest (and possibly only robust) fix is to call UpdateObserved()
for ServiceCompleteList and ignore Services entirely. This means that we
observe all services, even non-visible ones, but the extra overhead
should be negligible.

BUG=388423
R=armansito@chromium.org

Review URL: https://codereview.chromium.org/355573004