Export ResourceMessageParams::SerializedHandles
The inner class needs to be exported in case any uses in inline methods
or vtable of the outer class causes it to potentially be referenced across
a shared library boundary.

On non-Windows, the visibility attribute is already inherited by nested
classes, but on Windows, they must be marked explicitly.

BUG=82385
TEST=use Clang to do component release build of chrome on Windows

Review URL: https://codereview.chromium.org/359503003