Build ipclist as part of ipc_fuzzer and simplify .gyp files.
I was hoping that the msan tool would give us a dump of the uninit
messages it was finding, in which case ipclist will give a quick
back-mapping from number to msg class.  So ensure that it builds.

This will not build unless enable_ipc_fuzzer=1 is specified, so
should not impact build times.

R=inferno@chromium.org

Review URL: https://codereview.chromium.org/355203003