sync: Inheritance refactor for non-blocking sync
Implements several previously discussed changes to the inheritance
hierarchy for non-blocking sync classes.

Changes SyncContext from a wrapper around ModelTypeRegistry to an
interface class that ModelTypeRegistry implemnts.

Makes ModelTypeSyncWorkerImpl implement the ModelTypeSyncWorker
interface, even though it is never referenced via this interface.  Makes
the same change to ModelTypeSyncProxyImpl and the ModelTypeSyncProxy
interface.  These refactoring required that some methods be renamed to
be consistent with the interface.

This completes the refactoring that was started in r279618.

BUG=351005

Review URL: https://codereview.chromium.org/351073005