[Telemetry] Make our Enabled/Disabled annotations match what the bots run.
This change is not intended to add or remove anything from the bots. It is merely making
it so that 'run_benchmark list' will return the bot's current set of tests for
android.

This will enable us to just run all enabled benchmarks and add/remove benchmarks
without a master restart.

BUG=350973, 163503,388885

Review URL: https://codereview.chromium.org/335923007