Port History::QueryMostVisitedURLs to use CancelableTaskTracker
The callback no longer receive a HistoryService::Handle, but the client
should use the base::CancelableTaskTracker::TaskId to cancel a pending
call.

The callback receive a "const history::MostVisitedURLList*" as it is not
possible to pass a reference without adding a wrapper function. Since
there is a single client (apart from unit tests) this is the simplest
way to do it.

Merge HistoryBackend::QueryMostVisitedURLsImpl into the facade function
since they now have the same prototype.

BUG=371818

Review URL: https://codereview.chromium.org/352623002