Cast: increase the maximum outstanding audio frames
In the case of high latency congestion period, RTT is high and audio
frames are mostly delivered on time. We do no want to drop audio frames
frequently.

This change sends up to three times of target delay worth of audio
frames. It does not any harmful effect because audio packets are small
and receiver has the ability to drop any one of them.

Review URL: https://codereview.chromium.org/360103002