Don't use the frame that's going away in ProgrammaticScriptInjection
We're using a web frame that's being detached in ProgrammaticScriptInjection::
OnWillNotInject(), but it looks like this is leading to a crash (the frame
itself is probably still valid, but perhaps not all of its resources, e.g.
the url, document, etc). Just cache this so we don't rely on it.
BUG=390066

Review URL: https://codereview.chromium.org/336863010