WebRTC: Use same test name and display name for perf tests.
When the display name is different than the test name the
generated links to the build logs in the perf dashboard becomes
invalid. There are ways to resolve this by passing a 'a_stdio_uri'
value with the perf results, but it seems far eaiser to just
rename the test names for our tests.

BUG=382372
TEST=None

Review URL: https://codereview.chromium.org/333933008