Cleanup: Remove the second constructor of VariationsService.
This removes the constructor that takes just two parameters and do not
take |notifier| as it first parameter.

Since it is created by a factory method, we can get rid of this second
private constructor and just use the other constructor for both
production and tests code.

BUG=385843
TEST=None
R=asvitkine@chromium.org

Review URL: https://codereview.chromium.org/368523003