Reduces dependency from AutocompleteMatch to Profile (2nd).
The previous CL was reverted due to a crash in AutocompleteProviderTest.
That happened because RegisterTemplateURL utility in the test file
implicitly re-creates the TemplateURLService instance but this CL assumes
to keep the instance in a test case.

Therefore this CL changed the logic so that it does not re-create
TemplateURLService instance if it already exists.

BUG=388510
TBR=pkasting@chromium.org, dtrainor@chromium.org
TEST=unit_test --gtest_filter='*Autocomplete*.*'

Review URL: https://codereview.chromium.org/367923005