Do not create custom jumplists in tests
This prevents leakage of files in
~\AppData\Roaming\Microsoft\Windows\Recent\CustomDestinations
due to a side effect of browser navigations in tests.

BUG=389375
TEST=To verify that there are no regressions - on Win 7+, perform some
browser navigations and right click on chrome's taskbar icon. Verify
that the jumplist entries are updated accordingly.
To verify the fix - execute some browser tests and verify that no
*.customDestinations-ms files are added to the folder above.

Review URL: https://codereview.chromium.org/361823007