Don't call -makeKeyWindow when the omnibox gets focus.
Prior to this change, bubbles which bubbles which open as key window
but close on resign key close immediately after appearing, or as soon as
the omnibox gets key window.

The call to -makeKeyWindow was added to fix ctrl+l when a permission
bubble was visible.  It is no longer needed because the permission
bubble (which does not close on resign key) does not open as key
window.  If it doesn't open as key window, the ctrl+l event is
correctly handled by the browser window.

BUG=390880

Review URL: https://codereview.chromium.org/363103003