Call into ViewTargeter from View::HitTestRect()
Add a ViewTargeterDelegate member to ViewTargeter, which 
is a weak pointer back to the View on which the targeter 
is installed. Make RootView subclass ViewTargeterDelegate 
so that it can be installed as the delegate on its 
ViewTargeter.

Change the default implementation of View::HitTestRect()
to call into the ViewTargeter installed on |this| (or
instead on the ViewTargeter installed on the root view
if |this| does not have its own targeter installed).
This will allow all custom hit-testing logic to be moved
out of View::HitTestRect() overrides and into
ViewTargeterDelegate::DoesIntersectRect() overrides
instead.

BUG=388838
TBR=sky@chromium.org
TEST=ViewTargeterTest.DoesIntersectRect augmented, all existing tests involving hit testing

Review URL: https://codereview.chromium.org/360383002