CC: Clean up tile size setting.
This is a non-functional cleanup.

IMG introduced a memory leak so we switch to idle uploads.
The new drivers with the memory leak fixed don't need the
other workarounds either, so we can do without the size
work-arounds.

Without that work-around, we no longer need to know if the
driver is IMG, so we can move the tile-size logic to a
better place and avoid using the command-line.

BUG=365877

Review URL: https://codereview.chromium.org/313403002