Improve YouCompleteMe handling of Blink header without source files.
In this case, simply giving the Chromium src root as the include path
is unlikely to be helpful. We can get a pretty close approximation by
getting the flags for an arbitrary Blink source file, which will have
more reasonable include paths and defines for a Blink header.

Review URL: https://codereview.chromium.org/368313003