Pepper: Small PnaclResources cleanup
This fixes a place where the wrong URL is used for the ld process when
performing PNaCl translation. I don't believe the effect of this is severe,
though it will cause the wrong url to be displayed when it's running.

This also removes an unnecessary string constant usage in the trusted plugin.

BUG=

Review URL: https://codereview.chromium.org/373003005