MacViews: Consolidate kBrowserActionContainerWidth pref registration
This pref is currently registered in different places for Mac and
toolkit-views, resulting in problems for toolkit-views on Mac (i.e. it
gets registered twice).

The views pref moved into extension_prefs.cc in r183799. This moves the
mac pref to the same place for increased code cohesion, simplifying a
lot of things.

BUG=155525, 390755

Review URL: https://codereview.chromium.org/378153002