Impl thread smooth scrolling.
This is a first pass at smooth scrolling on the compositor thread.  It adds LayerTreeHostImpl::ScrollAnimated, which kicks off a scroll offset animation through the layer animation controller.

When the --enable-smooth-scrolling flag is present, InputHandlerProxy will call ScrollAnimated for composited wheel scrolls.

Once this is in the next steps are:

(1) Update the animation curve to aim at a new target when a wheel event comes in while a scroll animation is in progress.  (For now we just drop the event.)

(2) Unify the behavior with WebCore::ScrollAnimatorNone.  We should be able to share a lot of logic since cc::ScrollOffsetAnimationCurve is already exposed to Blink through compositor bindings (for http://crbug.com/243871).

BUG=575

Review URL: https://codereview.chromium.org/361143002