PasswordForm: move from current/old password scheme to current/new.
With the old approach (storing the current and old passwords for a form), we could not tell apart login forms from sign-up forms and change password forms that did not have a password confirmation field.

Furthermore, the new approach align more nicely with the new-password and current-password scheme we use for the autocomplete attributes.

BUG=375333

Review URL: https://codereview.chromium.org/356223002