Hotfix to make clang-by-default work with gcc 4.8
A better fix is to set clang further up, so that gcc_version is set to
0 when clang is on automatically
( https://codereview.chromium.org/384633005 ). This is however blocked
on nacl also using the "clang" symbol, so that needs to be fixed first
( https://codereview.chromium.org/384793002 ). To make the build go for
people in the meantime, land this hack. I'll revert it once the nacl change has
landed in nacl and then rolled into chromium, and the better fix has landed.

BUG=360311
TBR=thestig@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/379693003