Add components and CLD to GN build.
New components:
- rappor
- data_reduction_proxy
- captive_portal
- translate
- usb_service
- web_modal

Removes the _IMPLEMENTATION define for web_modal. It is unused (it is a static library, not a component).

Adds third_party/cld and third_party/cld_2 BUILD files.

Separates out the include dir setting of "//" and the root gen directory so. cld_2 needed to not have this due to base file name conflicts.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/382663002