Display json output for List Perf Tests
This patch put output of "List Perf Test" step in a placeholder provided by json
module, instead of leaking it out to a magic file.
In a later step that is expecting a json file as input, json module is again
used to put the data in a temp file before passing it to test_runner.py.
Such change makes the json output displayed on the builder page.

BUG=388885
R=tonyg@chromium.org

Review URL: https://codereview.chromium.org/383933002