Fix build dependencies on target remoting_webapp_unittest
There is a typo in the output element of the "Build Remoting Webapp unittest.html" action.  The current output element refers to a non-existent location.  As a result, ninja invoke the action on every build run.

I also removes exclude-js from the input element, as there is no need to re-invoke the action if the content of an excluded script is modified.

BUG=392834

Review URL: https://codereview.chromium.org/385713003