WebTestProxy should explode if readback requests are unhandled
Async readback requests are a potential use-after-free if the
WebTestProxy goes away before they are serviced.  However, this is not
actually anything a valid test would want to do if it wanted to not be
racy.  So, just CHECK for outstanding tasks on destruction rather than
posting with a weak pointer and silently failing such readbacks.

BUG=389569

Review URL: https://codereview.chromium.org/382093002