[Preferences] Change a preference from Integer to String
kContextualSearchEnabled should be a string instead of an integer, to be consistent with the calling code.

This pref was never set (ever!) on people's profiles, only registered in profile.cc. I verified with gab@ that moving to String at this point will not mess things up.

BUG=341761
R=noms@chromium.org

Review URL: https://codereview.chromium.org/386893003