Allow "cross-origin" navigations from about:blank in AreURLsInPageNavigation
This can happen when an iframe is opened, then popualted via a document.write()
from its parent. This will cause the url to change to the parent's url, but the
browser process will not be notified of this url change. If the iframe then
attempts a fragment navigation, it looks like a cross-origin navigation from
about:blank.

BUG=390798
TEST=Added case to NavigationControllerTest.IsInPageNavigation

Review URL: https://codereview.chromium.org/372403002