cc: Explicitly invalidate all dropped recordings on the main thread.
Currently PictureLayerTiling drops tiles when bounds change under the
assumption that these recordings have been dropped by the main thread,
however it misses some cases. For instance if the layer grows, and the
main thread has to drop recording tiles along the growing edge, but it
is outside the interest rect, then the PictureLayerTiling has no idea
that these recordings are gone, which leads to us having Tiles on the
compositor thread which can not be rastered.

Instead of having the impl side try to guess at these things, we now
have the PicturePile explicitly expand invalidation to cover every
recoding tile that has its picture dropped or that is removed from the
recording. Then the impl-side PictureLayerTiling only drops what the
main thread tells it to, and does not make any guesses of its own.

Perf sheriffs: This is expected to change results for "invalidation"
tests as it changes what code paths those tests execute.

R=enne
BUG=386998

Review URL: https://codereview.chromium.org/375923005