aura: Make sure redirected events have the correct location.
If an event is redirected to a different WindowTreeHost because the target
window belongs to that, then make sure the event is in the WindowTreeHost's
root-window's coordinate space. This is necessary to make sure that the
target (and the pre/post-target handlers) see the event in the correct
location.

BUG=385770
R=sky@chromium.org

Review URL: https://codereview.chromium.org/380343002